pyproject.toml
[tool.poetry]
name = "claficle"
version = "0.1.0"
description = "Cross-lingual adaptation for In Context Learning"
authors = ["Giulio Starace <giulio.starace@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
datasets = "^2.4"
transformers = "^4.21.1"
pytorch-lightning = "^1.7.0"
hydra-core = "^1.2.0"
torch = "^1.12.1"

[tool.poetry.dev-dependencies]
jupyterlab = "^3.4.3"
black = "^22.6.0"
flake8 = "^4.0.1"
jupyterlab-code-formatter = "^1.4.11"
ipykernel = "^6.15.0"
nbdime = "^3.1.1"
ipywidgets = "^7.7.1"
googletrans = "4.0.0rc1"
types-PyYAML = "^6.0.11"
mypy = "^0.971"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pyright]
reportPrivateImportUsage = false

[additional.notes]
bigbench = { url = "https://storage.googleapis.com/public_research_data/bigbench/bigbench-0.0.1.tar.gz" }
